+++ react/StorefrontOrganizationDetails.tsx
@@ -1,0 +1,3 @@
+import StorefrontOrganizationDetails from './components/OrganizationDetails'
+
+export default StorefrontOrganizationDetails